+++ internal/e2e-js/tests/callfabric/videoRoom.spec.ts
@@ -1,3 +1,4 @@
+import { uuid } from '@signalwire/core'
 import { Video } from '@signalwire/js'
 import { test, expect } from '../../fixtures'
 import {
@@ -5,17 +6,20 @@ import {
   createCFClient,
   expectLayoutChanged,
   expectMCUVisible,
+  getStats,
   setLayoutOnPage,
 } from '../../utils'
 
 test.describe('CallFabric VideoRoom', () => {
   test('should handle joining a room, perform actions and then leave the room', async ({
     createCustomPage,
+    resource,
   }) => {
     const page = await createCustomPage({ name: '[page]' })
     await page.goto(SERVER_URL)
 
-    const roomName = 'cf-e2e-test-room'
+    const roomName = `e2e-video-room_${uuid()}`
+    await resource.createVideoRoomResource(roomName)
 
     await createCFClient(page)
 
@@ -264,5 +268,55 @@ test.describe('CallFabric VideoRoom', () => {
     })
 
     expect(roomSession.success).toBe(false)
+  })
+
+  test('should handle joining a room with audio channel only', async ({
+    createCustomPage,
+    resource,
+  }) => {
+    const page = await createCustomPage({ name: '[page]' })
+    await page.goto(SERVER_URL)
+
+    const roomName = `e2e-video-room_${uuid()}`
+    await resource.createVideoRoomResource(roomName)
+
+    await createCFClient(page)
+
+    // Dial an address with audio only channel
+    const roomSession = await page.evaluate(
+      async ({ roomName }) => {
+        return new Promise<any>(async (resolve, _reject) => {
+          // @ts-expect-error
+          const client = window._client
+
+          const call = await client.dial({
+            to: `/public/${roomName}?channel=audio`,
+            rootElement: document.getElementById('rootElement'),
+          })
+
+          call.on('room.joined', resolve)
+          call.on('room.updated', () => {})
+
+          // @ts-expect-error
+          window._roomObj = call
+
+          await call.start()
+        })
+      },
+      { roomName }
+    )
+
+    expect(roomSession.room_session).toBeDefined()
+    expect(
+      roomSession.room_session.members.some(
+        (member: any) => member.member_id === roomSession.member_id
+      )
+    ).toBeTruthy()
+
+    const stats = await getStats(page)
+
+    expect(stats.outboundRTP).not.toHaveProperty('video')
+
+    expect(stats.inboundRTP.audio.packetsReceived).toBeGreaterThan(0)
   })
 })